post and response working fine

+++ src/components/Post/DeletePost.tsx
@@ -2,12 +2,18 @@ import { Button } from "@material-ui/core";
 import React, { Component } from "react";
 import {PostData} from '../Helpers/Interfaces';
 import APIURL from '../Helpers/environment';
-
+declare var window: any;
 
 type acceptedProps ={
     // updateToken: string | null;
+    data: {
+        id: number,
+        title: string,
+        description: string
+      } | any,
     updateToken: (newToken: string) => void,
     sessionToken: string | null,
+    onDone: () => void
 }
 
 type postData={
@@ -24,9 +30,12 @@ class DeletePost extends Component<acceptedProps, postData> {
     
     
     deletePost = (): any => {
+        if(!window.confirm('Are you sure you want to delete this item?')){
+            return;
+        }
         if(this.props.sessionToken !== null){
 
-            let id:number = 1; 
+            let id:number = this.props.data.id; 
             fetch(`${APIURL}/post/${id}`, {
                 method: 'DELETE',
                 headers: {
@@ -35,7 +44,7 @@ class DeletePost extends Component<acceptedProps, postData> {
                 },
             })
             .then(res => res.json())
-            .then(data => console.log(data))
+            .then(data => {console.log(data); if(typeof this.props.onDone === 'function') {this.props.onDone()} })
             .catch(err => console.log(err))
         }
     }
